+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.0.4
 - Preserve entry list scroll position in Manager when adding, deleting, or reordering entries
 - Save Manager scroll position per user and timeline, automatically restored on next view
+- Manager remembers last selected timeline and reopens it (if still exists)
 - Save viewer scroll position per user and timeline, automatically restored on next view
 - Fix Deprecated warning
 - Small CSS fix for renderJournalDirectory